Reorganise the intersection finding code
rename to a better name - it finds intersections, it doesn't check
them... Segment finding will have to work differently to path
intersection finding, so hive that off leaving space for an
IntersectionFinder::Path
