+++ lib/draught/intersection_finder.rb
@@ -1,0 +1,6 @@
+require_relative 'intersection_finder/segment'
+
+module Draught
+  module IntersectionFinder
+  end
+end
